requirements.txt
pip~=24.2
attrs~=23.2.0
colorlog~=6.8.2
wheel~=0.43.0
Pillow~=9.4.0
tornado~=6.4
Jinja2~=3.1.3
setuptools~=65.5.0
packaging~=23.0
pyparsing~=3.0.9
aiogram~=3.5.0
python-dotenv~=1.0.1
requests~=2.31.0